requirements.txt
# CLANK! Tracker - Screen capture and OCR for gameplay monitoring
pillow>=10.0.0          # Image processing
opencv-python>=4.8.0    # Computer vision and image processing
pytesseract>=0.3.10     # OCR text recognition
numpy>=1.24.0           # Numerical operations
matplotlib>=3.7.0       # Data visualization and plotting
pyautogui>=0.9.54       # Screen capture and automation
easyocr>=1.7.0          # Alternative OCR engine
requests>=2.31.0        # HTTP requests for API calls
pyyaml>=6.0             # Configuration file parsing
python-dotenv>=1.0.0    # Environment variable management

# Development dependencies
pytest>=7.4.0           # Testing framework
black>=23.0.0           # Code formatting
flake8>=6.0.0           # Code linting